+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.13.4] - 8/31/2021
+
+### Changed
+
+- Renamed `Optional.FirstOf` -> `Optional.FirstWithValue`
+
 ## [2.13.3] - 8/31/2021
 
 ### Added